requirements.txt
aliyun-python-sdk-core>=2.0.7
aliyun-python-sdk-alidns>=2.0.7
PyYAML>=3.12
validate_email>=1.3
#cryptography>=2.1.3
#crypto>=1.4.1
pyOpenSSL>=17.3.0
